fix: resolve all clippy warnings across workspace (#695)
- Add check-cfg for tarpaulin in Cargo.toml
- Fix unused_mut in state_backends.rs and template_functions.rs
- Replace deprecated IndexMap::remove with shift_remove in runtime.rs
- Fix String vs &str in gpu.rs run_cmd calls
- Fix provision.rs: add Path import, await WorkspaceManager::new
- Add structural clippy allows in lib.rs for project-wide patterns
- Fix MigrationOutcome imports in plan_parity.rs tests
- Fix test file API drift in parallel_stress, chaos, and docker tests
- Apply cargo clippy --fix across all targets

+++ src/native/users.rs
@@ -317,7 +317,7 @@ pub fn get_user_groups(username: &str) -> NativeResult<Vec<String>> {
         // Convert GIDs to group names
         let mut group_names = Vec::new();
         for gid in groups {
-            if let Ok(Some(group)) = get_group_by_gid(gid as u32) {
+            if let Ok(Some(group)) = get_group_by_gid(gid) {
                 group_names.push(group.name);
             }
         }
